Optimize post rendering hydration
